+++ db.go
@@ -41,6 +41,9 @@ var (
 
 	// ErrEntryIdxModeOpt is returned when set db EntryIdxMode option is wrong.
 	ErrEntryIdxModeOpt = errors.New("err EntryIdxMode option set")
+
+	// ErrFn is returned when fn is nil.
+	ErrFn = errors.New("err fn")
 )
 
 const (
@@ -175,11 +178,19 @@ func Open(opt Options) (*DB, error) {
 
 // Update executes a function within a managed read/write transaction.
 func (db *DB) Update(fn func(tx *Tx) error) error {
+	if fn == nil {
+		return ErrFn
+	}
+
 	return db.managed(true, fn)
 }
 
 // View executes a function within a managed read-only transaction.
 func (db *DB) View(fn func(tx *Tx) error) error {
+	if fn == nil {
+		return ErrFn
+	}
+
 	return db.managed(false, fn)
 }
 
@@ -208,7 +219,7 @@ func (db *DB) Merge() error {
 
 	for _, pendingMergeFId := range pendingMergeFIds {
 		off = 0
-		f, err := NewDataFile(db.getDataPath(int64(pendingMergeFId)), db.opt.SegmentSize)
+		f, err := NewDataFile(db.getDataPath(int64(pendingMergeFId)), db.opt.SegmentSize, db.opt.RWMode)
 		if err != nil {
 			db.isMerging = false
 			return err
@@ -242,18 +253,23 @@ func (db *DB) Merge() error {
 				if err == io.EOF {
 					break
 				}
+				f.rwManager.Close()
 				return fmt.Errorf("when merge operation build hintIndex readAt err: %s", err)
 			}
 		}
 
 		if err := db.reWriteData(pendingMergeEntries); err != nil {
+			f.rwManager.Close()
 			return err
 		}
 
 		if err := os.Remove(db.getDataPath(int64(pendingMergeFId))); err != nil {
 			db.isMerging = false
+			f.rwManager.Close()
 			return fmt.Errorf("when merge err: %s", err)
 		}
+
+		f.rwManager.Close()
 	}
 
 	return nil
@@ -282,6 +298,8 @@ func (db *DB) Close() error {
 
 	db.closed = true
 
+	db.ActiveFile.rwManager.Close()
+
 	db.ActiveFile = nil
 
 	db.BPTreeIdx = nil
@@ -292,7 +310,7 @@ func (db *DB) Close() error {
 // setActiveFile sets the ActiveFile (DataFile object).
 func (db *DB) setActiveFile() (err error) {
 	filepath := db.getDataPath(db.MaxFileID)
-	db.ActiveFile, err = NewDataFile(filepath, db.opt.SegmentSize)
+	db.ActiveFile, err = NewDataFile(filepath, db.opt.SegmentSize, db.opt.RWMode)
 	if err != nil {
 		return
 	}
@@ -323,7 +341,7 @@ func (db *DB) getMaxFileIDAndFileIDs() (maxFileID int64, dataFileIds []int) {
 		dataFileIds = append(dataFileIds, idVal)
 	}
 
-	sort.Sort(sort.IntSlice(dataFileIds))
+	sort.Ints(dataFileIds)
 	maxFileID = int64(dataFileIds[len(dataFileIds)-1])
 
 	return
@@ -365,8 +383,7 @@ func (db *DB) parseDataFiles(dataFileIds []int) (unconfirmedRecords []*Record, c
 	for _, dataID := range dataFileIds {
 		off = 0
 		fID := int64(dataID)
-		f, err := NewDataFile(db.getDataPath(fID), db.opt.SegmentSize)
-
+		f, err := NewDataFile(db.getDataPath(fID), db.opt.SegmentSize, db.opt.FileLoadingMode)
 		if err != nil {
 			return nil, nil, err
 		}
@@ -378,7 +395,7 @@ func (db *DB) parseDataFiles(dataFileIds []int) (unconfirmedRecords []*Record, c
 				}
 
 				e = nil
-				if db.opt.EntryIdxMode == HintAndRAMIdxMode {
+				if db.opt.EntryIdxMode == HintKeyValAndRAMIdxMode {
 					e = &Entry{
 						Key:   entry.Key,
 						Value: entry.Value,
@@ -410,10 +427,12 @@ func (db *DB) parseDataFiles(dataFileIds []int) (unconfirmedRecords []*Record, c
 				if off >= db.opt.SegmentSize {
 					break
 				}
-
+				f.rwManager.Close()
 				return nil, nil, fmt.Errorf("when build hintIndex readAt err: %s", err)
 			}
 		}
+
+		f.rwManager.Close()
 	}
 
 	return