+++ tx_zset.go
@@ -28,7 +28,7 @@ import (
 // SeparatorForZSetKey represents separator for zSet key.
 const SeparatorForZSetKey = "|"
 
-// ZAdd adds the specified member with the specified score and specified val to the sorted set stored at key.
+// ZAdd adds the specified member key with the specified score and specified val to the sorted set stored at bucket.
 func (tx *Tx) ZAdd(bucket string, key []byte, score float64, val []byte) error {
 	var buffer bytes.Buffer
 
@@ -45,7 +45,7 @@ func (tx *Tx) ZAdd(bucket string, key []byte, score float64, val []byte) error {
 	return tx.put(bucket, newKey, val, Persistent, DataZAddFlag, uint64(time.Now().Unix()), DataStructureSortedSet)
 }
 
-// ZMembers returns all the members of the set value stored at key.
+// ZMembers returns all the members of the set value stored at bucket.
 func (tx *Tx) ZMembers(bucket string) (map[string]*zset.SortedSetNode, error) {
 	if err := tx.checkTxIsClosed(); err != nil {
 		return nil, err
@@ -54,7 +54,7 @@ func (tx *Tx) ZMembers(bucket string) (map[string]*zset.SortedSetNode, error) {
 	return tx.db.SortedSetIdx[bucket].Dict, nil
 }
 
-// ZCard returns the sorted set cardinality (number of elements) of the sorted set stored at key.
+// ZCard returns the sorted set cardinality (number of elements) of the sorted set stored at bucket.
 func (tx *Tx) ZCard(bucket string) (int, error) {
 	members, err := tx.ZMembers(bucket)
 	if err != nil {
@@ -64,7 +64,7 @@ func (tx *Tx) ZCard(bucket string) (int, error) {
 	return len(members), nil
 }
 
-// ZCount returns the number of elements in the sorted set at key with a score between min and max and opts.
+// ZCount returns the number of elements in the sorted set at bucket with a score between min and max and opts.
 // opts includes the following parameters:
 // Limit        int  // limit the max nodes to return
 // ExcludeStart bool // exclude start value, so it search in interval (start, end] or (start, end)
@@ -78,7 +78,7 @@ func (tx *Tx) ZCount(bucket string, start, end float64, opts *zset.GetByScoreRan
 	return len(nodes), nil
 }
 
-// ZPopMax removes and returns up to count members with the highest scores in the sorted set stored at key.
+// ZPopMax removes and returns the member with the highest score in the sorted set stored at bucket.
 func (tx *Tx) ZPopMax(bucket string) (*zset.SortedSetNode, error) {
 	item, err := tx.ZPeekMax(bucket)
 	if err != nil {
@@ -88,7 +88,7 @@ func (tx *Tx) ZPopMax(bucket string) (*zset.SortedSetNode, error) {
 	return item, tx.put(bucket, []byte(" "), []byte(""), Persistent, DataZPopMaxFlag, uint64(time.Now().Unix()), DataStructureSortedSet)
 }
 
-// ZPopMin removes and returns up to count members with the lowest scores in the sorted set stored at key.
+// ZPopMin removes and returns the member with the lowest score in the sorted set stored at bucket.
 func (tx *Tx) ZPopMin(bucket string) (*zset.SortedSetNode, error) {
 	item, err := tx.ZPeekMin(bucket)
 	if err != nil {
@@ -98,7 +98,7 @@ func (tx *Tx) ZPopMin(bucket string) (*zset.SortedSetNode, error) {
 	return item, tx.put(bucket, []byte(" "), []byte(""), Persistent, DataZPopMinFlag, uint64(time.Now().Unix()), DataStructureSortedSet)
 }
 
-// ZPeekMax returns up to count members with the highest scores in the sorted set stored at key.
+// ZPeekMax returns the member with the highest score in the sorted set stored at bucket.
 func (tx *Tx) ZPeekMax(bucket string) (*zset.SortedSetNode, error) {
 	if err := tx.checkTxIsClosed(); err != nil {
 		return nil, err
@@ -111,7 +111,7 @@ func (tx *Tx) ZPeekMax(bucket string) (*zset.SortedSetNode, error) {
 	return tx.db.SortedSetIdx[bucket].PeekMax(), nil
 }
 
-// ZPeekMin returns up to count members with the lowest scores in the sorted set stored at key.
+// ZPeekMin returns the member with the lowest score in the sorted set stored at bucket.
 func (tx *Tx) ZPeekMin(bucket string) (*zset.SortedSetNode, error) {
 	if err := tx.checkTxIsClosed(); err != nil {
 		return nil, err
@@ -124,7 +124,7 @@ func (tx *Tx) ZPeekMin(bucket string) (*zset.SortedSetNode, error) {
 	return tx.db.SortedSetIdx[bucket].PeekMin(), nil
 }
 
-// ZRangeByScore returns all the elements in the sorted set at key with a score between min and max.
+// ZRangeByScore returns all the elements in the sorted set at bucket with a score between min and max.
 func (tx *Tx) ZRangeByScore(bucket string, start, end float64, opts *zset.GetByScoreRangeOptions) ([]*zset.SortedSetNode, error) {
 	if err := tx.checkTxIsClosed(); err != nil {
 		return nil, err
@@ -137,7 +137,7 @@ func (tx *Tx) ZRangeByScore(bucket string, start, end float64, opts *zset.GetByS
 	return tx.db.SortedSetIdx[bucket].GetByScoreRange(zset.SCORE(start), zset.SCORE(end), opts), nil
 }
 
-// ZRangeByRank returns all the elements in the sorted set in one bucket at bucket and key
+// ZRangeByRank returns all the elements in the sorted set in one bucket and key
 // with a rank between start and end (including elements with rank equal to start or end).
 func (tx *Tx) ZRangeByRank(bucket string, start, end int) ([]*zset.SortedSetNode, error) {
 	if err := tx.checkTxIsClosed(); err != nil {